requirements.txt
Flask==2.2.5
PyAudio==0.2.13
PyYAML==6.0
requests==2.31.0
openai